.github: add a reusable azure vm infra setup for build & test
Add a reusable Github Actions workflow for setting up Azure vm
to run unit tests. This workflow is intended to be called using
`workflow_call` from other workflows. It is used to provision the
environment before executing tests on it.

Add a workflow to build and run unit tests for the mshv crate.
It calls the infra setup workflow to provision then logs in via
SSH and runs the tests. It ensures tests run in a consistent VM
setup.

Signed-off-by: Aastha Rawat <aastharawat@microsoft.com>